publisher: move kafka to dedicated package
