+++ publisher/kafka/kafka_test.go
@@ -1,4 +1,4 @@
-package publisher
+package kafka
 
 import (
 	"fmt"
@@ -8,6 +8,7 @@ import (
 	"github.com/confluentinc/confluent-kafka-go/kafka"
 	"github.com/raystack/raccoon/logger"
 	pb "github.com/raystack/raccoon/proto"
+	"github.com/raystack/raccoon/publisher"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 )
@@ -31,7 +32,7 @@ func TestProducer_Close(suite *testing.T) {
 		client := &mockClient{}
 		client.On("Flush", 10).Return(0)
 		client.On("Close").Return()
-		kp := NewKafkaFromClient(client, 10, "%s", 1)
+		kp := NewFromClient(client, 10, "%s", 1)
 		kp.Close()
 		client.AssertExpectations(t)
 	})
@@ -55,7 +56,7 @@ func TestKafka_ProduceBulk(suite *testing.T) {
 					}
 				}()
 			})
-			kp := NewKafkaFromClient(client, 10, "%s", 1)
+			kp := NewFromClient(client, 10, "%s", 1)
 
 			err := kp.ProduceBulk([]*pb.Event{{EventBytes: []byte{}, Type: topic}, {EventBytes: []byte{}, Type: topic}}, group1)
 			assert.NoError(t, err)
@@ -79,22 +80,22 @@ func TestKafka_ProduceBulk(suite *testing.T) {
 				}()
 			}).Once()
 			client.On("Produce", mock.Anything, mock.Anything).Return(fmt.Errorf("buffer full")).Once()
-			kp := NewKafkaFromClient(client, 10, "%s", 1)
+			kp := NewFromClient(client, 10, "%s", 1)
 
 			err := kp.ProduceBulk([]*pb.Event{{EventBytes: []byte{}, Type: topic}, {EventBytes: []byte{}, Type: topic}, {EventBytes: []byte{}, Type: topic}}, group1)
-			assert.Len(t, err.(BulkError).Errors, 3)
-			assert.Error(t, err.(BulkError).Errors[0])
-			assert.Empty(t, err.(BulkError).Errors[1])
-			assert.Error(t, err.(BulkError).Errors[2])
+			assert.Len(t, err.(*publisher.BulkError).Errors, 3)
+			assert.Error(t, err.(*publisher.BulkError).Errors[0])
+			assert.Empty(t, err.(*publisher.BulkError).Errors[1])
+			assert.Error(t, err.(*publisher.BulkError).Errors[2])
 		})
 
 		t.Run("Should return topic name when unknown topic is returned", func(t *testing.T) {
 			client := &mockClient{}
 			client.On("Produce", mock.Anything, mock.Anything).Return(fmt.Errorf("Local: Unknown topic")).Once()
-			kp := NewKafkaFromClient(client, 10, "%s", 1)
+			kp := NewFromClient(client, 10, "%s", 1)
 
 			err := kp.ProduceBulk([]*pb.Event{{EventBytes: []byte{}, Type: topic}}, "group1")
-			assert.EqualError(t, err.(BulkError).Errors[0], "Local: Unknown topic "+topic)
+			assert.EqualError(t, err.(*publisher.BulkError).Errors[0], "Local: Unknown topic "+topic)
 		})
 	})
 
@@ -115,13 +116,13 @@ func TestKafka_ProduceBulk(suite *testing.T) {
 					}
 				}()
 			}).Once()
-			kp := NewKafkaFromClient(client, 10, "%s", 1)
+			kp := NewFromClient(client, 10, "%s", 1)
 
 			err := kp.ProduceBulk([]*pb.Event{{EventBytes: []byte{}, Type: topic}, {EventBytes: []byte{}, Type: topic}}, "group1")
 			assert.NotEmpty(t, err)
-			assert.Len(t, err.(BulkError).Errors, 2)
-			assert.Equal(t, "buffer full", err.(BulkError).Errors[0].Error())
-			assert.Equal(t, "timeout", err.(BulkError).Errors[1].Error())
+			assert.Len(t, err.(*publisher.BulkError).Errors, 2)
+			assert.Equal(t, "buffer full", err.(*publisher.BulkError).Errors[0].Error())
+			assert.Equal(t, "timeout", err.(*publisher.BulkError).Errors[1].Error())
 		})
 	})
 }
