+++ publisher/kafka/mockclient.go
@@ -1,4 +1,4 @@
-package publisher
+package kafka
 
 import (
 	"github.com/confluentinc/confluent-kafka-go/kafka"
