+++ publisher/pubsub/pubsub.go
@@ -105,7 +105,7 @@ func (p *Publisher) ProduceBulk(events []*pb.Event, connGroup string) error {
 	}
 
 	if cmp.Or(errors...) != nil {
-		return publisher.BulkError{Errors: errors}
+		return &publisher.BulkError{Errors: errors}
 	}
 	return nil
 }
